tests/Document-createProcessingInstruction-literal-2.xhtml

<?xml-stylesheet href="support/style.css" type="text/css"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Document.createProcessingInstruction</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com"/>
<link rel="help" href="../web-dom-core#processinginstruction"/>
<script src="testharness.js"></script>
</head>
<body>
<div id="log"/>
<script>
test(function() {
  var pienc = document.firstChild;
  assert_true(pienc instanceof ProcessingInstruction)
  assert_equals(pienc.target, "xml-stylesheet")
  assert_equals(pienc.data, 'href="support/style.css" type="text/css"')
})
</script>
</body>
</html>
